// src/hooks/useHeader/useHeader.enum.ts
// Enum for the header sections in English
export enum HeaderSectionEnglishEnum {
  HOME = 'Home',
  HOTEL_MANAGEMENT = 'Hotel management',
  INVESTORS = 'Investors',
  TECHNOLOGY = 'Technology',
  RESTAURANTS = 'Restaurants',
  ACCESS = 'Access',
  LEADERSHIP_TEAM = 'Leadership Team',
  CONTACT = 'Contact',
  LANGUAGES = 'Languages',
  EXPERIENCE = 'Experience',

}

// Enum for the header sections in Spanish
export enum HeaderSectionSpanishEnum {
  HOME = 'Home',
  HOTEL_MANAGEMENT = 'Gestion hotelera',
  INVESTORS = 'Inversionistas',
  TECHNOLOGY = 'Tecnologia',
  RESTAURANTS = 'Restaurantes',
  ACCESS = 'Access',
  LEADERSHIP_TEAM = 'Lideres',
  CONTACT = 'Contacto',
  LANGUAGES = 'Lenguajes',
  EXPERIENCE = 'Experiencia',
}